refactor(dashboard): extract KeeperActivitySummary and split keeper-detail-panels tests
- Extract inline runtime activity summary from keeper-detail-body.ts
  into KeeperActivitySummary component (keeper-detail-activity-summary.ts)
- Reduces keeper-detail-body.ts from 371 → 329 lines
- Split monolithic keeper-detail-panels.test.ts into 3 focused files:
  - keeper-detail-ctx-utils.test.ts (autonomyHint, formatDuration)
  - keeper-detail-lists.test.ts (RelationshipList, TraitsList)
  - keeper-detail-telemetry.test.ts (InferenceTelemetryPanel)
