+++ dashboard/src/components/keeper-detail-telemetry.test.ts
@@ -1,63 +1,12 @@
 import { render } from 'preact'
 import { html } from 'htm/preact'
 import { describe, it, expect, afterEach } from 'vitest'
-import type { Keeper, KeeperMetricPoint} from '../types'
-import { autonomyHint, formatDuration } from './keeper-detail-ctx-utils'
-import {
-  InferenceTelemetryPanel,
-  RelationshipList,
-  TraitsList,
-} from './keeper-detail-panels'
+import type { Keeper, KeeperMetricPoint } from '../types'
+import { InferenceTelemetryPanel } from './keeper-detail-telemetry'
 
 afterEach(() => {
   document.body.innerHTML = ''
 })
-
-describe('autonomyHint', () => {
-  it('returns active hint when count is 0 and proactive enabled', () => {
-    expect(autonomyHint(0, true)).toBe('활성 · 미발동')
-  })
-
-  it('returns disabled hint when count is 0 and proactive disabled', () => {
-    expect(autonomyHint(0, false)).toBe('자율 비활성')
-  })
-
-  it('returns disabled hint when count is 0 and proactive undefined', () => {
-    expect(autonomyHint(0, undefined)).toBe('자율 비활성')
-  })
-
-  it('returns undefined when count is positive', () => {
-    expect(autonomyHint(5, true)).toBeUndefined()
-    expect(autonomyHint(1, false)).toBeUndefined()
-  })
-
-  it('returns disabled hint when count is undefined', () => {
-    expect(autonomyHint(undefined, undefined)).toBe('자율 비활성')
-  })
-})
-
-// ── formatDuration ────────────────────────────────────────────
-
-describe('formatDuration', () => {
-  it('formats seconds under 60', () => {
-    expect(formatDuration(0)).toBe('0초')
-    expect(formatDuration(30)).toBe('30초')
-    expect(formatDuration(59)).toBe('59초')
-  })
-
-  it('formats minutes under 3600', () => {
-    expect(formatDuration(60)).toBe('1분')
-    expect(formatDuration(120)).toBe('2분')
-    expect(formatDuration(3599)).toBe('59분')
-  })
-
-  it('formats hours with remaining minutes', () => {
-    expect(formatDuration(3600)).toBe('1시간 0분')
-    expect(formatDuration(3660)).toBe('1시간 1분')
-    expect(formatDuration(7384)).toBe('2시간 3분')
-  })
-})
-
 
 describe('InferenceTelemetryPanel', () => {
   it('separates wall tok/s from hardware decode tok/s', () => {
@@ -251,31 +200,5 @@ describe('InferenceTelemetryPanel', () => {
     expect(container.textContent).toContain('API latency')
     expect(container.textContent).not.toContain('0.0s')
     expect(container.querySelector('svg[aria-label="API 지연 시간 추이"] polyline')).toBeNull()
-  })
-})
-
-describe('RelationshipList and TraitsList primitives', () => {
-  it('renders relationship names with the shared status chip primitive', () => {
-    const container = document.createElement('div')
-    document.body.appendChild(container)
-
-    render(html`<${RelationshipList} rels=${{ alpha: 'mentor' }} />`, container)
-
-    const chip = container.querySelector('[data-status-chip]')
-    expect(chip?.textContent).toContain('alpha')
-    expect(chip?.getAttribute('data-status-chip-tone')).toBe('info')
-    expect(chip?.getAttribute('data-status-chip-uppercase')).toBe('false')
-  })
-
-  it('renders trait labels with the shared status chip primitive', () => {
-    const container = document.createElement('div')
-    document.body.appendChild(container)
-
-    render(html`<${TraitsList} label="traits" traits=${['planner']} />`, container)
-
-    const chip = container.querySelector('[data-status-chip]')
-    expect(chip?.textContent).toContain('planner')
-    expect(chip?.getAttribute('data-status-chip-tone')).toBe('info')
-    expect(chip?.getAttribute('data-status-chip-uppercase')).toBe('false')
   })
 })
